Improve screen jumping algorithm
The current algorithm is roughly:
1. For each monitor in the same order the OS numbers them
    1. If the cursor moved off the current monitor horizontally (including if it also moved vertically, i.e. diagonally), and the monitor's edge lines up exactly with the current monitor, jump to that monitor
    2. if the cursor moved of the current monitor vertically, and the monitor's edge lines up exactly with the current monitor, jump to that monitor

This causes errors, as shown in issue #16, when more than one monitor is in the direction the cursor leaves the screen. Instead we'd like to pick the "best" screen in the direction the cursor leaves the screen. "Best" isn't well defined, but closest by euclidean distance should be a reasonable approximation.

The new algorithm is:
1. Filter the list of all monitors to just those in the direction the cursor left the current monitor.
2. Jump to the monitor in the filtered list which is closest to where the cursor left the current monitor.

Fixes #16