Remove "Rescale pointer between screens" option
Since this feature was never completed, has unintuitive behavior, and there doesn't seem to be much demand for it, remove it. If we ever want to add a version of this feature, it won't take much effort to re-write it from scratch. See issue #8 for more complete discussion on its behavior and the decision to remove it.

Closes #8